feat(yuhyeon): Phase 04 — 일시정지 메뉴 (부분 완료, 시각 검증 보류)
코드/구조/InputSystem/Inspector 연결 100% 완료. ESC 콜백 동작 확인
(Hierarchy MenuRoot 토글). 단 Game 뷰 시각 렌더링 미통과 — 다음
세션에서 Unity MCP 활성 + Debug.Log 출력으로 30초 안에 진단 예정.

산출물:
- Scripts/UI/PauseMenuController.cs — Dawnholder.Client.UI 네임스페이스,
  InputActionReference 패턴, Toggle + 3 버튼 핸들러, Time.timeScale 처리,
  헌법 #1 (Server Authority) 주석 명시
- Scenes/UI.unity — PauseMenuCanvas (Sort Order 20) + MenuRoot(풀스크린
  stretch, 시작 OFF) + Backdrop + MenuPanel(중앙 300x240, Vertical Layout)
  + ResumeButton/MainMenuButton/QuitButton (LayoutElement Preferred Height 60)
- InputSystem_Actions — UI Action Map에 TogglePauseMenu 액션 추가
  (Binding: <Keyboard>/escape, Type: Button)
- Phase 04 정의 사후 정정 — PauseMenuCanvas 배치 씬을 Gameplay → UI로
  변경 (ADR-021 영역 분리 정책 일관성, 정의 작성 후 결정된 사항)
- Phase 04 -DONE.md 박제 (부분 완료 명시 + 학습 보존 + 다음 세션 액션)

부수 fix (팀장 영역 임시 패치, PR에 명시):
- Dawnholder.Client.asmdef references에 Unity.TextMeshPro 추가 (영구 fix)
  → 본인 영역 Scripts/UI/HudController.cs (TMP_Text) 컴파일 회귀 복구
  → 팀장 PR(M2 Phase 06, cee8775) asmdef 신설 시 누락된 참조
- LocalPlayerController.cs:64 — (int)_localTickCounter 캐스트 추가
- UnityClientSession.cs:145 — (uint)pkt.lastAckedClientTick 캐스트 추가
  → 두 캐스트 모두 TEMP-yuhyeon-20260517 마커 + 주석 (PDL 재생성 시 제거)
  → 원인: tick counter uint 통일 결정 후 팀장이 PDL 재생성 누락한 채
     main push (CHANGELOG 미박제) → pull 받자마자 빌드 회귀

학습 보존 (★★★ /journal:bug 후보):
- "Unity UI 손작업의 함정 — 코드는 정상인데 조용히 안 보이는 디버깅 지옥"
- 코드/InputSystem/Inspector 다 정상이지만 RectTransform/Anchor/Canvas
  Scaler/Color α/SetActive/Sort Order 조합 중 어딘가 깨지면 *조용히* 안 보임
- MCP/스크립트 자동화의 가치를 직접 체감한 사건

+++ 03_Client/Assets/Scripts/Input/LocalPlayerController.cs
@@ -61,7 +61,10 @@ namespace Dawnholder.Client.Input
             C_MoveIntent pkt = new C_MoveIntent
             {
                 inputX = encoded,
-                clientTick = _localTickCounter
+                // TEMP-yuhyeon-20260517: PDL의 clientTick이 int인데 _localTickCounter는 uint —
+                // tick counter uint 통일 결정(learning-journal/youngho/int-vs-uint-for-tick-counters.md)
+                // 후 팀장이 PDL 재생성 빼먹은 채 main push → 빌드 깨짐. PDL이 uint로 재생성되면 캐스트 제거.
+                clientTick = (int)_localTickCounter
             };
             // Phase 05: SendIntent 경유 — Editor에서 SimulatedLatencyMs 적용 가능.
             session.SendIntent(pkt.Write());
